1.0

+++ datasets/dataloader_assd.py
@@ -82,10 +82,11 @@ class ASSD(Dataset):
             cropper = T.RandomCrop(size=(int(4000*p), int(6000*p)))
 
             img_gt = cropper(torch.cat((img, gt), 0))
-            img, gt = img_gt[:3, :, :], img_gt[3, :, :].unsqueeze(0).long()
+            # To uniform output format
+            img, gt = img_gt[:3, :, :], img_gt[3, :, :].unsqueeze(0)
 
         img = self.normalize(self.shrink(img))
-        gt = self.shrink(gt).squeeze(0)
+        gt = self.shrink(gt.long()).squeeze(0)
         return img, gt
